Move API to the new class
